% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{changeLabels}
\alias{changeLabels}
\title{Change Label of data columns}
\usage{
changeLabels(graphicData, dataType, myLabel)
}
\arguments{
\item{graphicData}{data for each summary graphic}

\item{dataType}{The type of each variable}

\item{myLabel}{Labels to replace variable name with}
}
\value{
the same tibble but with changed labels
}
\description{
Change Label of data columns
}
\examples{
2+2
}
