% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{cutData}
\alias{cutData}
\title{cutData
cut appropriate columns of the tibble (ignoring the first two which should be rownumber and id)}
\usage{
cutData(graphicData, dataTypeList, graphicStyle, graphic)
}
\arguments{
\item{graphicData}{data for each summary graphic}

\item{dataTypeList}{the data type within each graphicData tibble}

\item{graphicStyle}{the style of graphic to be produced (typically default)}

\item{graphic}{the exact type of graphic (one of table or plot)}
}
\value{
returns a tibble with cut data as formatted factors
}
\description{
cutData
cut appropriate columns of the tibble (ignoring the first two which should be rownumber and id)
}
\examples{
2+2
}
