% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingREDCapData.R
\name{expandDataDictionary}
\alias{expandDataDictionary}
\title{Expand the data dictionary to account for checkbox entries}
\usage{
expandDataDictionary(dataDictionary)
}
\arguments{
\item{dataDictionary}{the data dictionary which describes the data set}
}
\value{
a tibble similar to data dictionary but with formatted checkbox entries
}
\description{
Expand the data dictionary to account for checkbox entries
}
\examples{
'none'
}
