% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureGen.R
\name{figureGen}
\alias{figureGen}
\title{generate figures for plotting}
\usage{
figureGen(graphicData, graphicStyle, graphic, dataType, subsection)
}
\arguments{
\item{graphicData}{Data to generate figure from (table or plot or custom)}

\item{graphicStyle}{A string specifying what type of subgraphic to generate}

\item{graphic}{A string specifing table, plot or custom.}

\item{dataType}{A sting specifying the data types found in graphicData}

\item{subsection}{a string specifying the subsection these graphics will be found in}
}
\value{
A figure of the type specified
}
\description{
generate figures for plotting
}
\examples{
'TBD'
}
