% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{getData}
\alias{getData}
\title{getDaTa
Return appropriate data sections from a REDCap style database}
\usage{
getData(REDCapdata, variableName, instrumentLabel)
}
\arguments{
\item{REDCapdata}{A REDcap style tibble}

\item{variableName}{The variable(s) to extract from the dataframe}

\item{instrumentLabel}{The instrument to extract the data from}
}
\value{
A tibble of the selected data
}
\description{
getDaTa
Return appropriate data sections from a REDCap style database
}
\examples{
'1+1'
}
