% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingREDCapData.R
\name{readDataDictionary}
\alias{readDataDictionary}
\title{Load a modified version of a REDCap Data dictionary}
\usage{
readDataDictionary(path, delim = ",")
}
\arguments{
\item{path}{a path to a modified version of the redcap data dictionary as a csv.}

\item{delim}{the deliminator for the file type}
}
\value{
a formatted data dictionary
}
\description{
Load a modified version of a REDCap Data dictionary
}
\examples{
'none'
}
