% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingREDCapData.R
\name{readREDCapData}
\alias{readREDCapData}
\title{readREDCapData}
\usage{
readREDCapData(path, longerDataDictionary, delim = ",")
}
\arguments{
\item{path}{a path to the redcap data file}

\item{longerDataDictionary}{a data dictionary preferably one which has been
made longer to include the subvariables for checkboxes}

\item{delim}{the deliminator used such as ',' for .csv files}
}
\value{
a formatted REDCAP data base as a tibble
}
\description{
readREDCapData
}
\examples{
'none'
}
