% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapDataSummary-package.R
\docType{package}
\name{REDCapDataSummary-package}
\alias{REDCapDataSummary}
\alias{REDCapDataSummary-package}
\title{REDCapDataSummary: Create data summaries from REDCap}
\description{
This package is intended to aid the reading of data from a REDCap database and then produce a summary which can produced in a document from rmarkdown to aid in the process of data cleaning.
}
\author{
\strong{Maintainer}: Robert Liddell \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
